% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ps.R
\name{common_ps}
\alias{common_ps}
\title{common_ps}
\usage{
common_ps(ps, numCores, colname)
}
\arguments{
\item{ps}{A phylyseq object}

\item{colname}{Name of column of grouping variable, as a string}

\item{numcores}{An integer with number of cores to use}
}
\value{
A lsubset phlyoseq object
}
\description{
A function that calls common_taxa to return a new ps object with only common taxa
}
\examples{
    common_ps(small, 2, "KitType")

}
