% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_taxa.R, R/unique_taxa.R
\name{unique_taxa}
\alias{unique_taxa}
\title{common_taxa}
\usage{
unique_taxa(ps)

unique_taxa(ps)
}
\arguments{
\item{ps}{A phylyseq object}

\item{numcores}{An integer with number of cores to use}

\item{colname}{Name of column of grouping variable, as a string}
}
\value{
A list of strings

A list of strings
}
\description{
A function which returns a list of asvs/column names found in common in a group.

A function which returns the column names of a ps object as a list
}
\examples{
    common_taxa(small, 2, "KitType")

}
